[FIX] documents_account: Fix versioning invoice main attachment
Steps to reproduce:
-------------------

- Install `documents_account` module
- Go to settings -> activate 'Centralize accounting files and documents'
- Still in settings, under centralization feature, click on Journals to
  create a journal to synchronize
- Create a record with Vendor Bills as journal and with any workspace
- Go to account module and create a vendor bill in draft
- Attach 2 files at once on the vendor bill
- Go to Document module, open the new document and replace the
  attachment (versioning feature)
- Go back to the bill and switch the 'Main Attachment'
- Go back to document and restore versioned attachment
- Go back to the bill and switch again the 'Main Attachment
- Go back to document

Issue:
------

A second document is created.

Cause:
------

The issue is when switching the second time the invoice main attachment,
the current main attachment is in fact in the versioned attachments
(`document.preview_attachment_ids`), and since we search only for
document that have as attachment (`document.attachment_id`) the
current invoice main attachment, we don't find it and create a new
document.

Solution:
---------

Look also in the versioned attachments when searching for the invoice
main attachment in case it has been versioned.

opw-4028789

closes odoo/enterprise#67507

X-original-commit: a4194cb84a1934d1fd41f9fc8b64ee8e70fc4d82
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Nasreddin Boulif (bon) <bon@odoo.com>