[FIX] account_reports: Map annotations instead of filter them
The aim of this commit is changing the way we're getting annotations
for a specific line.
Before, we had to iterate on each annotation to know if we can keep it
or not by using the related line ID.
Now, visible annotations are a mapping `{line ID: annotations}`,
so we just have to select annotations by using the lineID.
It avoids several filters in the code.

task-4045835

closes odoo/enterprise#67782

X-original-commit: 1133c665890c980b13c75175f33dc1ffd05ce5c3
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>