[PERF] mrp_workorder: make shop floor blazing fast
Speed up the shop floor by drastically reducing the number of refresh
calls.
We try to avoid a refresh of data from the server as much as possible.
If an update is still unavoidable, we try to only refresh the relevant
MO.
The data in the view now also gets refreshed automatically if no manual
refresh was performed by the user in the last 10 minutes.

task-4048415

closes odoo/enterprise#67498

X-original-commit: 791d2c76baf1bf47db6abb190fe759ce7862326b
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Claeys Pieter (clpi) <clpi@odoo.com>