[FIX] knowledge: block rename attempt if no edition rights
This commit fixes an issue with locked/readonly empty articles.
If a user opens an article which is locked/readonly for them, when switching
articles the browser will try to update the article's name.
This operation is of course not possible, but this leads to the user not able to
leave said article.
This is caused by the update method on the record which sets it as dirty so that
it can be saved when possible.

In order to fix, this when the user has no write access on the article or if it
is locked, we return early so that the record doesn't set itself as dirty.

task-4047722

closes odoo/enterprise#67564

X-original-commit: 41d3af7cc95ba6024e47c75875bea751ac5eda86
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>