[FIX] sale_subscription: prevent type errror on handling failed transaction
Currently, an error occurs when the payment status is failed, expired, or failed.

Steps to produce an error:
- Install `sale_subscription` and `payment_mollie`
- setup `mollie` payment provider in testing
- Create an invoice for the subscription and confirm it
- Click preview of the invoice and click Pay Now.
- fill card details, and on the on the next page, select failed or
  expired in payment status
- Confirm

error: 
`PaymentTransaction._set_canceled() takes 1 positional argument but 2 were given`

This is because at code [1], we have passed `**kwargs` as arguments in super
call, and as a result, it sets all arguments of kwargs in agruments, e.g., 2 of
super call, but method call requires one positional argument.

This commit will fix the above issue by passing a positional argument with
kwargs in the super call.

[1]-https://github.com/odoo/enterprise/blob/43621084d0cf95970eb6e45cf67bcb20ec14b8b3/sale_subscription/models/payment_transaction.py#L186

sentry-5606090812

closes odoo/enterprise#66664

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>