[FIX] l10n_mx_edi: fix negative base amount in CFDI with down payment
Steps to reproduce:
- Install Accounting, Sales, l10n_mx_edi
- Switch to a Mexican company (e.g. ESCUELA KEMPER URGATE)
- Create a SO:
  * Customer: [a Mexican customer] (e.g. INMOBILIARIA CVA)
  * Order Lines:

       Product   |   Quantity   |   Unit Price   |   Taxes
    ---------------------------------------------------------
      Product A  |     1.00     |     1000.00    |    16%
      Product B  |     1.00     |     1500.00    |    16%
      Product C  |     1.00     |     3000.00    |    16%

- Confirm the SO
- Create a 90% down payment
- Confirm down payment and generate CFDI via "Send & Print" button
- Create a regular invoice for the remaining amount
- Confirm the invoice and generate CFDI

Issue:
The verification of the CDFI fails because there is one line with
a negative base amount.

Cause:
The down payment line has a negative amount.
This negative amount is dispatched between the other lines and the
"base" and "importe" values of these other lines are recomputed
depending on the amount that is deducted from it.
However, the computation is based on a ratio computed from the price subtotal
of the negative line, but the ratio is then applied to the remaining "base"
amount of the negative line, which decreases for each deducted line.
This generates incorrect values in "transferred_values_list" field for the
the second line and the following ones on which some amount is dispatched.
The issue only happens on the last line on which some amount has been
dispatched because it receives the remaining amount that is too high due to
the values deducted by the other lines being lower than what it should has
been.

Solution:
The ratio should be computed from the remaining "base" amount of the negative
line.

opw-4033979

closes odoo/enterprise#66992

X-original-commit: 2e995c018ed166281e2913f499b21cf5b50f5f38
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>