[FIX] pos_restaurant_appointment: manage bookings crash
Steps to reproduce:

1. Install pos_restaurant_appointment.
2. Start a restaurant session.
3. Find and click the "Manage Bookings" button from the top-right burger menu.
   -> Whole app crashes.

This is because the button is trying to open a dialog containing a custom gantt
view (`appointment_booking_gantt`) that is not loaded when opening pos.

This commit proposes a fix by including the custom gantt view in the pos assets
to prevent this crash and to fully utilize the booking management by the
custom gantt view.

closes odoo/enterprise#66731

X-original-commit: 9860f7b4a1900f734b902479d21a5ad5417083bd
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>