[IMP] hr_contract_salary: change behavior of salary configurator link
Previously, if some values were changed on the offer and then the salary configurator link
was clicked without saving the offer, the changes weren't reflected on the configurator
because the offer hadn't been saved yet.

These changes address this issue by changing the behavior of the salary configurator link.
Now, when the link is clicked, the offer is saved and the user is redirected to the
configurator webpage.

task-4008437

closes odoo/enterprise#65529

Related: odoo/upgrade#6221
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>