[FIX] account_budget: remove js_class='budget_list_view'
With https://github.com/odoo/odoo/commit/f1749d3299957e2949e0b78653266a5397bb213c, if a js_class is found in the arch but not in the view
registry, we no longer fallback to the basic view type and a crash
occurs. Here we simply remove a js_class from an arch to avoid such a
crash.

Build 65543538

closes odoo/enterprise#66744

X-original-commit: 367ca2ceb486e7432cde71daaea2870fd20c9bab
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>