[REF] web_tour: prepare to remove in_modal in step tour structure
In order to simplify the structure of a tour step, it was decided to
remove the "in_modal" key. The purpose of this key is to search for
the trigger in a modal element. But actually you just need to add
".modal" to the selector.
This functionality therefore really has little added value.
That's why we're removing it.
In this commit, we prepare the ground to be able to remove this
functionality by adding .modal to the selectors and forcing the fact
of not looking in a modal (in_modal: false) otherwise that would
be duplicative.

task~3974087

closes odoo/enterprise#66131

Related: odoo/odoo#172081
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>