[FIX] spreadsheet_dashboard_hr_payroll
- Go to Dashboards > Configuration > Dashboards
- Open the Human Resources section
- Open the payroll dashboard
- Open the Period filter configuration panel
- the field matching period offset of pivot 3 and 4 are
  inverted.

In the Data sheet, the column named "current" (pivot 3) is filtering on
previous, and the one called "previous" (pivot 4) is not filtering on anything.

Note that this fix is only active on new databases or if the dashboard is
deleted then the module "spreadsheet_dashboard_hr_payroll" is updated.

closes odoo/enterprise#66760

Task: 4056412
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>