[REF] l10n_mx_edi_stock: merge mx_edi_stock_extended with mx_edi_stock
To use the mexican delivery EDI, it is now required that we have the
address field from the mx_edi_extended (the module that stock_extended
previously additionally depends on). Using base mx_edi_stock without the
extended module will not work anymore.

With that in mind, it is now a good time to make the base stock depends
on extended and merge everything from the stock extended to the base
stock module.

closes odoo/enterprise#67304

Task-id: 3894340
Signed-off-by: John Laterre (jol) <jol@odoo.com>