[FIX] l10n_mx_edi_stock_30: Fix traceback on batch pickings install
When clicking on Batch Transfers in the settings to install
stock_picking_batch on a database having l10n_mx_edi_stock_30 installed
a traceback is raised due to the field l10n_mx_edi_is_delivery_guide_needed
which is missing as it is removed in stock_picking_batch.

Steps to reproduce:
- Install l10n_mx_edi_stock_30
- In Inventory -> Configuration -> Settings
- Check 'Batch Transfers', and Save

A ParseError traceback appears.

This commits aims to solve the issue by adding the field before its
parent element is removed from the arch.

closes odoo/enterprise#67105

X-original-commit: 1ca5029ab99aede1b6967242c69f782a8042c74a
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Alexandre Claes (alcl) <alcl@odoo.com>