[FIX] web_studio: apply datetime widget when unticking showSeconds
After making changes in community PR: https://github.com/odoo/odoo/pull/162373

The datetime widget was not being applied when we dragged and dropped a datetime
field from the studio.

Steps to reproduce:

- Open the Sales module and toggle Studio.
- Drag and drop a datetime field into the list view.
- Try to tick and untick the showSeconds option.

Observed behavior:
When ticking and unticking the showSeconds option, the seconds are not hiding
before resetting the widget.

Expected behavior:
The seconds should be hidden when we tick and untick the showSeconds option
without resetting the widget.

After this commit, the datetime widget will be applied when we drag and drop a
datetime field from the studio and untick the showSeconds boolean.

Task-2517675

closes odoo/enterprise#61553

Related: odoo/odoo#162373
Related: odoo/documentation#10326
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>