[FIX] web_studio: reportEditor: fix undeterministic tour failures
Before this commit, in about 8% of the time for some test tours, some edition of the report
was triggered before the OdooEditor had a chance to start, so the concerned step failed.

After this commit, there is no such race condition, as we add OdooEditor's specific class to the
step selectors.

runbot-error-29949 (and all linked error)

closes odoo/enterprise#66815

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>