[FIX] l10n_fr_reports: neutralize
This commit adds the missing neutralization necessary for the
l10n_fr_reports module introduced in
06f32096749a8df29b248f29698226cf996e9106

The purpose of the standard neutralization framework is to allow us to
create database copies that will not interact with external systems in
ways that could impact the production database (or if it is not possible
to prevent the interactions, make sure that they are benign or won't
result in actual changes), or impact the customers of the operator of
the production database.

This is mainly useful to allow safe support investigation on database
duplicates.

opw-3928199

closes odoo/enterprise#66838

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>