[FIX] website_knowledge: use the correct user for ACL test
This commit changes the user for the test
`:TestWKnowledgeSecurity.test_models_as_user` which should test ACLs for
internal users.
Initially it used `portal_user`, which is not internal but portal.
It is now replaced by `employee`.

We also update the values for asserts on stages as the number of accessible
articles for `employee` is different.
This user has access to more articles which may contain stages.

task-4059967

closes odoo/enterprise#66994

X-original-commit: 51e22958726773fad9e780d5ed3ef700084e5be4
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>