[FIX] sale_renting: ensure rental period is visible when creating a new ro
Version: 17.0
Steps:
Click on the schedule feature in the rental module.
Click on the 'New' to create the rental order.

Issue :
When a user clicks on the "Schedule" feature in the rental module and then
clicks on "New" to create a rental order, the rental period feature is disabled,
and the payment period appears instead.

Cause :
The issue is due to it not get the value in context of order_id.

fix:
To fix this issue, we add 'in_rental_app': True into the context.
This ensures that the rental period feature is always visible when creating a
new rental order.

opw-4016866

closes odoo/enterprise#69534

X-original-commit: b9a4c5a66f0d75284c9a818389898761f8187632
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>