[FIX] account_followup: pdf does not have the lines of the branches.
To reproduce:
Create a branch for your company
On this branch, create an overdue invoice for a partner.
Change company for the parent company.
Go to Follow-up Reports
Go on this partner.
You see the invoice appearing in the list
Click on Follow-up and print
=> The pdf does not contain the invoice from the branch

The issue is that we check on company and not on the company and branches.

opw-3988003

closes odoo/enterprise#67033

X-original-commit: 51fd2f4333e635e9218f830c63dd7523147b459a
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>