[FIX] whatsapp: composer sample values not updated on template change
Before this Commit
===============
In WhatsApp Composer, the Sample Values do not update when changing
the WhatsApp Template. Instead, it display the sample values from the
initially selected template.

This issue also occurs with the Button URL as well.

After this Commit
==============
The Sample Values for Free Text Fields and Button Dynamic URL will
now update based on the selected WhatsApp Template.

Task-3996935

closes odoo/enterprise#67416

X-original-commit: 6f1b1a843314d383c4e5ee369fba1ef6d68fdb75
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>