[IMP] hr_payroll: Allow using inputs in attachment if no matching code
Purpose
=======

This behavior was there to avoid applying a child allowance on a payslip
if no rules had the same code than the salary attachment type, to prevent
counting the salary attachment as applied if the amount wasn't retrieved
on the payslips.

Now that salary attachment can be used in a flexible way, this doesn't make
sense anymore.

Instead, you should configure properly the "available in structures" field
on the salary input type itself.

Part-of: odoo/enterprise#66638
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>