[FIX] *: dark mode for lazy views
In https://github.com/odoo/odoo/commit/f1749d3299957e2949e0b78653266a5397bb213c, we forgot to prepare a bundle for the dark mode. As a
consequence, the grid view does not have the style it should have in dark
mode. Here we use the newly introduced bundle 'web.assets_backend_lazy_dark'
in https://github.com/odoo/odoo/pull/173620 to fix that and make studio
use that bundle too in dark mode.
In this forward port we also fix the style of the gantt view in dark mode.

closes odoo/enterprise#67200

X-original-commit: 9b638cf7a2b7b05b27c22204b1fe9afed7c9e153
Related: odoo/odoo#174238
Signed-off-by: Pierre Paridans (app) <app@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>