[FIX] account_reports: better fiscal position search
Description of the issue this commit addresses:

During dev of the new annotation system, the way the fiscal position was
included in the annotations domain was by fetching all the amls of the report
that were linked to a fiscal position and getting the fiscal positions of the
moves linked to those amls. This can mean fetching a lot of amls only for a few
or even none fiscal positions.

Desired behavior after this commit is merged:

Performance are improved by avoiding searching the account.move.line model.
This commit relies on the way fiscal positions are attributed on the creation
of the annotations to get the right fiscal positions when loading a report.

task-4045835

X-original-commit: 7c3a498dbffa64a7f83f7626b4b80e0e309268c7
Part-of: odoo/enterprise#68197
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>