[IMP] sale_ebay: match changed urls to variants
eBay recently changed the URL they send in the viewitemURL. We found 3 different ways they send it,
but there might be more. We shouldn't have relied on this URL to match products in the beginning,
but we can't refactor this flow in stable, instead, we try a different way to match the url, and
fallback on another eBay call to get the expected variant.

opw-3934127

closes odoo/enterprise#67449

X-original-commit: 3f6143583a2b6f979c4bc703f3f333a1d818d0a2
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>