[FIX] l10n_co_edi: exclude self-withholding taxes from totals
Previously, we added support for self-withholding taxes,
which do not need to be reported in the XML file.
Hence, tax lines that contain a tax with a code ZZ (for customer invoices)/ ZZ, 07 (for vendor bills) are filtered out.

However, when we compute the total amounts these are not filtered out, leading to a discrepancy in sums.

We should filter out such tax lines when computing totals as well.

See other commit:
https://github.com/odoo/enterprise/commit/af8f364cc1f09cc1f9159b326a7f59c6524380be

opw-3958518

closes odoo/enterprise#67038

X-original-commit: ccf1fe4d328d277f494f96a3a69ef3417809906d
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>