[IMP] website_sale_renting,*: accordion in product page
*: website_sale_subscription

Follow the changes made in community[1], this commit shows the rental
pricing table inside the product details accordion if the accordion is
enabled by the user. It additionnaly slightly adapts the rental pricing
outside of accordion to make it inline with the new styling.

The same is done for the subscription pricing table.

task-3987039
part of task-3986921

[1]: https://github.com/odoo/odoo/pull/173823

closes odoo/enterprise#67008

Related: odoo/odoo#173823
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>