[FIX] l10n_mx_edi_stock_30: fixed memory limit issue on installation
Installation of l10n_mx_edi_stock_30 module can fail due to memory limit Having a large number of "stock.picking" records causes this due to the computations of the initial values of fields.

Solving this requires creating the field columns and assigning the values manually, by overriding the _auto_init method. This allows us to skip the computations, to avoid the memory limit.

opw-3846355

closes odoo/enterprise#65830

Forward-port-of: #60948
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>