[FIX] l10n_mx_edi: Fix unreachable commit when success
- Create 4 invoices: A, B, C, D with a correct MX setup
- Make sure C will fail by removing the SAT code on the product for example
- Run the send and print wizard on all invoices and check 'Download' (to force a synchronous run).

After each call to the web-service, the code does:

```
                # Check for success.
                if invoice.l10n_mx_edi_cfdi_state == 'sent':
                    continue
```

However, the `cr.commit` is at the end of the loop. So the transaction is not committed except in case of error!
When running the send and print synchronously, the error is raised to the user that will rollback the whole
transaction.
In case of the setup above, A, B and C are well commit but D is sent and has been rollback.
Then, the user will sign again D making the invoice signed twice.

To solve the issue, let's commit right after each web-service call.
That way, whatever the way we use to call the web-services, the transaction is commit immediately.

opw-4003043
opw-3985376
opw-4059666
opw-3955553
...and many others!

closes odoo/enterprise#67488

X-original-commit: b111f86396d0325ab5a392485cded47c1e966b13
Related: odoo/odoo#174755
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>