[FIX] helpdesk: fix hoot tests + avoid unneeded rpc call in team form
Before this commit, the test checking the reload_context is called only
when the user enables for the first time a feature in helpdesk team.
The test `reload the page when the feature use_timesheet is enabled in all teams`
fails indeterminately because we don't wait enough to reload the page
once the user clicks on the arrow right of the pager to get the second
record. Moreover, it is not really possible to reproduce a hard reload in
HOOT test (unmount and mount a view) and so this test has to be reviewed.

This commit changes some conditions to call `check_modules_to_install`
only when a feature is enabled on helpdesk team and it also reviews the
test failed and updates the tests but also alter some tests to use hooks
developed in web module.

runbot-70509

closes odoo/enterprise#68017

X-original-commit: c8bc3e31efce6883095194c6eaf244bc7aee557c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>