[IMP] stock_barcode: continuous scanner
Before this commit, when the user uses the device camera to scan a
barcode, the camera video was opened in a dialog and once a barcode was
scanned, the dialog was closed, making the process laborious for the
user if they want to scan multiple barcodes one after the other.
With this commit, the camera stay opened in the app, letting the user to
scan barcodes until they close the camera.

The system parameter `stock_barcode.delay_between_scan` let the user
configure the delay between each scan. If not set, 2000 ms will be used
by default.

The input to manually entered barcode was moved in the operation's
actions menu.

task-4058325

Part-of: odoo/enterprise#63727
Related: odoo/odoo#167942
Related: odoo/upgrade#6249
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>