[FIX] l10n_employment_hero: fix lock date
Currently there is an `@api.onchange` function on 'res.company'.
It is supposed to ensure that the `employment_hero_lock_date` is not after the `fiscalyear_lock_date`.
But this does not work at all since both lock dates are changed in different views / wizards.

In this commit:
  - The broken onchange is removed.
  - A constraint on `employment_hero_lock_date` is added
  - A function is added to compute the needed date from the `employment_hero_lock_date` and the other relevant fiscal lock dates.

On the `employment_hero_lock_date` constraint:
It ensure that we do not set a Employment Hero Lock Date before the relevant fiscal Lock Dates.
This avoids unexpected behavior / user frustration; since we only start fetching from the maximum of both of these dates.
We accept that the Employment Hero Lock Date can get outdated in case we change the relevant fiscal Lock Dates:
The field is assumed to be mostly used when setting up the DB to fetch historical data.

task-3891414 (related; rework of the lock dates)

Part-of: odoo/enterprise#67090
Related: odoo/odoo#174004
Related: odoo/upgrade#6295
Signed-off-by: William André (wan) <wan@odoo.com>