[FIX] sale_account_accountant: fix matching special sequences
Currently, if a user uses a sale order sequence that contains special characters like `/` (e.g. SO/2024/12/31/01),
they do not see such a sale order in suggestions in bank reconciliation
if the label doesn't match the name of the sale order exactly.
That is, a label `SO/2024/12/31/01 test` would not much this sale order,
but a label `SO2024123101 test` would much a sale order with a name `SO2024123101`.

This is because we format tokens and filter out special characters and then compare them to names.

With this commit, we'll compare names and tokens that are formatted the same way.

opw-3964898

closes odoo/enterprise#67544

X-original-commit: 2da9133810d0eff27fca07fce51ad3d3c3b2d39f
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>
Co-authored-by: smetl <las@odoo.com>