[FIX] account_reports: fiscal position filtering
Description of the issue/behavior this commit fixes:

Specific instances of fiscal positions using country groups instead of
countries, we can have fiscal positions that should be displayed in the
Tax Report which end up not being shown and selectable.

Desired behavior after the commit is merged:

This commit fixes the FP that are excluded from
the fiscal position selector.

task-3088020

closes odoo/enterprise#67072

X-original-commit: 95dc8a17fa935ca0be147434b53d9389fa8510e4
Related: odoo/odoo#173985
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>