[FIX] account_reports: fiscal position filtering
Description of the issue/behavior this commit fixes:

Specific instances of fiscal positions using country groups instead of
countries, we can have fiscal positions that should be displayed in the
Tax Report which end up not being shown and selectable.

Desired behavior after the commit is merged:

This commit fixes the FP that are excluded from
the fiscal position selector.

task-3088020

closes odoo/enterprise#67065

X-original-commit: 49e0b1724cd186693ba3b08de65f581eb7b1706d
Related: odoo/odoo#173976
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>