[FIX] website_studio: fix main tour at name_slugified step
in tour engine, we listen for mutations in the DOM in order to look for
the element pointed to by the "trigger". If it is found, we execute
run().
In this tour, the div[name='name_slugified'] element is found even
before the textContent modifications have been made. To resolve the
problem, simply be more precise in the trigger selector.
It should be noted that we can therefore remove the first assertion in
run() which amounts to doing exactly the same thing as the trigger.

closes odoo/enterprise#67080

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>