[FIX] web_studio: reportEditor: do not interfere when adding editable nodes
Add a Banner in a report.
Before this commit, the banner's content was not editable. This was because the reportEditor
resetted the attribute contenteditable when adding a new node.

After this commit, the banner is editable. After the adapatation to the new Editor's architecture
we don't interfere in general with contenteditable and let that to the editor.

 Please enter the commit message for your changes. Lines starting

closes odoo/enterprise#67088

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>