[IMP] account_reports: move account code column to the right in xlsx
Currently, when exporting an accounting report as XLSX file, we add a
column for the account codes on the very left of the sheet. This looks
a bit weird and is not the most important information to have as a first
column.

In this commit, we move the column to the right so the name of the line
will be first, followed by the account code (if applicable). We also add
a column name for the account codes.

task-3986483

closes odoo/enterprise#67111

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>