[FIX] industry_fsm_sale: test_qty_to_invoice_from_fsm w/o stock
When stock is not installed, the override of _fsm_ensure_sale_order in
industry_fsm_stock doesn't auto-confirm sale order, so the test
`test_qty_to_invoice_from_fsm` failed because the `qty_to_invoice` was
set to 0 if the order was not confirmed.

With this PR, the order is confirmed in the test flow if stock is not
installed.

opw-4061485

closes odoo/enterprise#67219

X-original-commit: 033dfea15e12b708560d514d2ae9dd7509b42223
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>