[FIX] hr_payroll: fix input salary rule with only input condition
Issue: If you create a salary rule with a condition based on an input but with
something else than input for the amount, the rule will not be shown.

Expected: The rule should be shown if the input is present in the payslip and
the computation should happen normally.

Fix: When checking if the rule should be shown or not, we were looking at the
wrong field that is related to the amount and not the condition.

closes odoo/enterprise#67115

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>