[FIX] l10n_be_reports_post_wizard: fix country check
Currently we just check whether the company has 'BE' fiscal country code.
But this can be incorrect:
We may want to create a report for a foreign VAT fiscal position.
This situation can lead to a traceback (see below).

After this commit we check the country code of the report
(instead of the country code of the active company).

Reproduce:
1. Install `l10n_be_reports` and `l10n_fr_reports`
2. On the Belgian company create a fiscal position with country 'France' and a 'Foreign Tax ID'
3. On the Belgian company open the tax report and select the French tax report
4. Click on the "Closing Entry" button to create a draft closing entry move
5. Try to post the move
6. Traceback (in which `l10n_be_reports_post_wizard` appears)

task-4063968

closes odoo/enterprise#67827

X-original-commit: eb5d6a285b6f0942b4a6d34755cc7f49ea22e6fb
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>