[FIX] account_followup: make the Exclude field not autosave
This bug is only present in v17.0.

Steps to reproduce:
- Go in Accounting > Customers > Follow-up Reports
- Click on a customer
- Changing the "Exclude from Follow-up" field of a line
- The save button does not work anymore

Cause:
The field "Exclude from Follow-up" is in autosave. So it is saved everytime it is changed. When clicking on the save button there is nothing to save so nothing happens and the buttons stays there.

Solution:
Add `{options='autosave': False}` in the field to prevent it from saving. This will create the same behavior as in all other versions.

This bug does not appear in 17.1 because a new condition called canSaveOnUpdate is added: https://github.com/odoo/odoo/blob/2db71d627fd2c9fc97d573df7993a953753bb8bd/addons/web/static/src/model/relational_model/record.js#L256

opw-4055350

closes odoo/enterprise#67125

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>