[FIX] l10n_mx_edi : CFDI payment cancellation
This commit change the way payments are cancelled through cfdi. Now the cancellation reason and the cancellation process are automated and can be managed directly in the invoice view.

It also adds a new field in the "pay" wizard, where the user can mention the CFDI origin in case it's a cancellation with replacement.

task-3894308

closes odoo/enterprise#67630

Related: odoo/odoo#175015
Signed-off-by: Laurent Smet (las) <las@odoo.com>