[FIX] web_studio: prevent crash when no OuterGroup
Steps to reproduce
==================

- Install account_accountant,web_studio
- Go to Accounting > Configuration > Account Groups
- Open studio
- Switch to the form view
- Show invisible elements
- Add a new field before the invisible Company field

=> Error: Xpath resolved to nothing or multiple nodes

Cause of the issue
==================

The xpath props of the StudioHook is the string "null"

The fields of this view are defined inside an InnerGroup but there is no
OuterGroup.

```xml
<form><sheet>
    <group>
        <field .../>
        ...
    </group>
</sheet></form>
```

In the form compiler, when a group does not contain another group, it is
an InnerGroup

The form editor compiler uses some `interestingSelector` to compute xpaths.

It assumed that an InnerGroup was always inside an OuterGroup

Solution
========

Since the `:has` selector has not been widely supported for long enough,
we do the equivalent of `:not(field) group:not(:has(> group)) > *` in
javascript to match any group that has no direct group as a child.

opw-4046558

closes odoo/enterprise#67750

X-original-commit: 2b430e4aa689f70eb5391010c666a8b7e741d85d
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>