[IMP] l10n_br_edi: import official NF-e vendor bill XMLs
This adds support for importing electronically invoiced vendor bills
for goods. We follow the official Brazilian specification [1]. Service
bills are out of scope because each city has its own format.

This populates the bill with data that already exists in the database,
the only exception is that the vendor will be automatically created if
it doesn't yet exist.

task-3860901

[1] Esquema XML NF-e/NFC-e - Pacote de Liberação nº 9o on
    https://www.nfe.fazenda.gov.br/portal/listaConteudo.asp

closes odoo/enterprise#67573

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>