[FIX] sale_purchase_{stock, inter_company_rules}: transfer custom attribute values on SO
Steps to reproduce:

Have 2 companies: CO1 and CO2

In the setting:
- Enable Multi-Step Routes and Inter-Company Transactions
- With CO1 and CO2 Synchronize Sales and Purchase Order
- Go to Inventory > Configuration > Warehouse Management > Routes
- Unarchive the MTO Route
- With CO1 create a route with a buy action and destination CO1/stock
- With CO2 create a route with a manufacture action
- Go to Inventory > Configuration > Products > Attributes
- Create a product attribute with a value that "is_custom"
- Create a storable product P with using MTO and tour CO1 and CO2 routes
- Set CO2 as a vendor for the product
- With CO1, create a SO for 1 unit of your product P
- a popup asks you for a custom product attribute value, write "test"
- Confirm the SO > Automatically creates a PO (w/ description)
- Confirm the PO > Automatically create an SO in CO2 (w/ description)
- Confirm the SO in CO2 > Automatically create an MO

Issue:

The custom description is missing on the final MO.

Note:

If you were to create an SO directly in CO2 for 1 unit of P, the custom
description associated with the custom attribute value would be present
on the associated MO.

Cause of the issue:

The MO is created by a procurement generated from the SOL during the
call of the `_action_launch_stock_rule` method:
https://github.com/odoo/odoo/blob/055184ea033e7068ce33c92390c73ae39b2db259/addons/sale_stock/models/sale_order_line.py#L345
The custom description that will appear on the MO and containing the
informations related to the custom attribute values is generated here:
https://github.com/odoo/odoo/blob/055184ea033e7068ce33c92390c73ae39b2db259/addons/sale_stock/models/sale_order_line.py#L267
However, this method will return an empty string because the
`product_custom_attribute_value_ids` is absent from our SOL
https://github.com/odoo/odoo/blob/055184ea033e7068ce33c92390c73ae39b2db259/addons/sale/models/sale_order_line.py#L338-L347
On the other hand, this `product_custom_attribute_value_ids` was set
and is still existing on the original SO that started the flow in CO1.
The reason that it is not present anymore on the last SO is because this
SOL field has no equivalent on the `purchase.order.line` model and the
information was therefore forgotten on the MO and has not been
transferred directly from the PO of CO1 to the SO of CO2.

Fix:

We rely on the `sale_line_id` field of the `purchase.order.line` model
in order to determine the original SOL of the final SOL and update
the `product_custom_attribute_value_ids` accordingly.

Note:

As a side effect of our change in `sale_purchase_stock` linking our sol
to the pol via the procurement commit 63ef74b can be reverted since the
message is already automatically posted (the message is now written by
the user who cancelled the SO rather than Odoobot).

opw-3998861

closes odoo/enterprise#68288

X-original-commit: f9921d85663f142afa7e4517319a39275339b577
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>