[IMP] sign: disable "Add on all pages" button for single-page PDFs
Before this commit, when uploading a single-page PDF and dropping the "initial" field,
the wizard still offered the "Add on all pages" option, which is not useful for a one-page document.

After this commit, the "Add on all pages" button will be disabled if the document has only one page.
Also in case if there is only page we directly add field into pdf without hassle to choose signer.

task-4069435

closes odoo/enterprise#67237

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>