[FIX] appointment: fix traceback when computing leave resource intervals
Steps to reproduce
==================
1). Create an event in any appointment type
2). Remove the start date from it
-> Traceback occurs
video ref - ﻿﻿https://youtu.be/5NpfiKNdjbc

Technical
==========
in interval_from_events function when the end time or start time is not defined
then it will not be able to unpack the values to their appropriate variables due
to this error occur.

After this commit
=================
Now traceback will not occur.

Task-4010880

X-original-commit: 91d7066116f269e62f819af5a8638960c9753243
Part-of: odoo/enterprise#69221
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Shaan Thakkar (thsh) <thsh@odoo.com>