[FIX] hr_payroll_account, l10n_{}_hr_payroll_account: avoid failure at installation
Problem
---------
Currently, if you install l10n_xx and update some specific account codes
(depending on the localization) and then try to install the
corresponding l10n_xx_hr_payroll_account hr module, you are struck with
an error that cancels the installation as some accounts are missing.

1. Install l10n_ae for example
2. Switch to AE company
3. Update the account with code 201002 to 701002
4. Install l10n_ae_hr_payroll_account
-> Error occurs during the installation

Objective
---------
Be able to install the module even when the account is missing.

Solution
---------
Log a warning instead of raising an error.

opw-3961798

closes odoo/enterprise#67408

X-original-commit: 48c7c60319f7af69a60835fcbcb7944eaac022fb
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>