[FIX] l10n_es_reports: support pro rata tax in libros export
This commit implements the functionality to support exporting libros
XLSX with some line using a tax of type pro rata.

Pro rata is a fiscal system here in Spain, in which companies who have
activities where they can deduct IVA, and activities where they don´t
have that right, the fiscal Agency gives you a formula you can use to
simplify the process, and deduct the percentage you get from that
formula from the IVA amount of an invoice. For example, on an invoice
with tax 21% IVA, from that amount we can apply the pro rata percentage,
and get a deductable and non deductable amount. The deductable amount
should go to the normal IVA account, and the non deductable should go to
the same account as the tax base.

To simulate pro rata tax in Odoo, we can create a new tax with only
partial of the tax percentage is applied an account, tax grid, and
checked in 'use in tax closing'. When we create invoice/bill with this
tax, only those partial amount should be included in the tax_deductible
field of the libros XLSX.

Before this commit: tax_deductible amount include the full tax amount.

After this commit: tax_deductible amount only include the tax amount if
the line have a tax grid.

closes odoo/enterprise#67797

Task-id: 3955233
X-original-commit: c00d7085457026a34ec6980306d5c2576eb679cf
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>