[PERF] web_studio: Speedup view normalization in edit_view
When changes are made through studio, the view needs to be normalized first which in turn requires the use of `difflib.ndiff` that is notorious for slowness in some cases. Refer to this issue on `difflib` https://github.com/python/cpython/issues/119105 for an analysis of why it is slow. In our case, we only care about added and removed lines, not replaced lines, which means `ndiff` is not needed. Instead `unified_diff` is used which is much faster.

This PR adapts the module to use `difflib.unified_diff` over `difflib.ndiff` to speedup `edit_view` requests when making any modifications to the view.

|no. diff lines|before | after|
------------------|-----------|------------
|~4670 | ~1.4 minutes | ~2.5 sec|

opw-4054294

closes odoo/enterprise#77737

X-original-commit: 36c3e521eae4563dd91f2af826588cb7df0aa321
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Mohamed Kasem (kasm) <kasm@odoo.com>