[IMP] l10n_in_asset, account_asset: Improve indian degressive depreciation
according to the Company Act of 2013, depreciation should be calculated on the full cost 'original_value'
and the residual value should be 5%, which is 5,000 in this casefor the indian case only
so, it's not the general case. So, we need to overwrite the logic of the degressive depreication
to adapt the desired flow.

Now, We calculate the degressive amount by 'factor * current_value of asset' excluding 'salvage_value'

task-id#3909619

closes odoo/enterprise#68283

X-original-commit: e2b87fa6f688d9689662748ee2b1cb26f128a53c
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Muhamed Abdalla (muab) <muab@odoo.com>